% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contagion.R
\name{contagion}
\alias{contagion}
\title{Contagion Simulations}
\usage{
contagion(exposures, 
          buffer, 
          shock = "all", 
          weights = NULL, 
          method = "general", 
          ...,
          exposure_type = c("assets", "liabilities", "impact", "vulnerability"),
          keep.history = FALSE, 
          abs.tol = .Machine$double.eps ^ 0.2, 
          max.it = min(1000, nrow(v)*10),
          verbose = TRUE)
}
\arguments{
\item{exposures}{an adjacency \code{\link{matrix}}, (sparse) \code{\link{Matrix}} or an \code{\link[igraph]{igraph}}
object with the network of bilateral exposures between vertices. By default, the function
expects the exposures in the form of an assets matrix
in which A -> B means that A has an asset with B. However, you can
change that with the parameter \code{exposure_type}. When using a matrix, preferably it should have
rows and columns names.}

\item{buffer}{a numeric vector with the capital buffer for each vertex.
Values should be in the same row/column order as the network of bilateral exposures. The
buffer is not needed if \code{exposure_type = "vulnerability"}.}

\item{shock}{a list with the shock vectors. If \code{"all"} (the default) the function will run a 
simulation for the default of each vertex in the network.}

\item{weights}{default is \code{NULL}. You can use a numeric
vector of weights to give some economic significance to the measures, like,
for instance, the total assets of the nodes.}

\item{method}{the contagion propagation method.}

\item{...}{other arguments to be passed to the contagion propagation method.}

\item{exposure_type}{character vector indicating the type of the bilateral exposures. It can be
an \code{"assets"} network (where A -> B means that A has an asset with B),
a \code{"liabilities"} network (where A -> B means that A has a debt with B),
a (binary) \code{"impact"} matrix (where A -> B indicates the relative impact
 of A in B's capital buffer), or 
a (binary) \code{"vulnerability"} matrix 
(where A -> B indicates the relative impact A suffers from B's default). 
The default is \code{"assets"}.}

\item{keep.history}{keep all the history of stress levels? This can use a lot of memory, so
the default is \code{FALSE}.}

\item{abs.tol}{the desired accuracy.}

\item{max.it}{the maximum number of iterations.}

\item{verbose}{gives verbose output. Default is \code{TRUE}.}
}
\description{
Given a matrix of exposures, a vector of buffers and weights (optional) 
the functions simulates contagion for all the shock vectors provided. You may
choose from the implemented propagation contagion method or create you own propagation method.
}

