% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DebtRank.R
\name{debt_rank}
\alias{debt_rank}
\alias{debt_rank_shock}
\title{DebtRank of the vertices}
\usage{
debt_rank(exposures, capital_buffer, weights = rep(1, nrow(exposures)),
  binary = FALSE, exposure_type = c("assets", "liabilities",
  "vulnerability"), max.it = 100, abs.tol = 1e-09)

debt_rank_shock(exposures, capital_buffer, shock_vector, weights = rep(1,
  nrow(exposures)), binary = FALSE, exposure_type = c("assets",
  "liabilities", "vulnerability"), max.it = 100, abs.tol = 1e-09)
}
\arguments{
\item{exposures}{an adjacency \code{\link{matrix}} or an \code{\link[igraph]{igraph}}
object with the network of bilateral exposures between vertices. By default, the function
expects the exposures in the form of an assets matrix
in which A -> B means that A has an asset with B. However, you can
change that with the parameter \code{exposure_type}. When using a matrix, preferably it should have
rows and columns names.}

\item{capital_buffer}{a numeric vector with the capital buffer for each vertex.
Values should be in the same row/column order as the network of bilateral exposures. This
parameter is not needed if \code{exposure_type = "vulnerability"}.}

\item{weights}{default is equal weights to all vertices. This should be a vector
representing the relative importance of each vertex in the network.
If the vector does not sum to 1, it will be normalized. You should use
weights that have some practical significance like,
for instance, the total assets of the nodes in the case of a financial network.}

\item{binary}{if \code{binary = TRUE} it computes the binary vulnerability matrix,
truncating all values less than 1 to 0.}

\item{exposure_type}{character vector indicating the type of the bilateral exposures. It can be
an "assets" network (where A -> B means that A has an asset with B),
a "liabilities" network (where A -> B means that A has a debt with B) or
it can be a (binary) "vulnerability" matrix (where A -> B indicates the relative impact
 of A in B's capital buffer).For the last option see \code{\link{vulnerability_matrix}}.
 The default is "assets".}

\item{max.it}{the maximum number of iterations. Default is 100.}

\item{abs.tol}{the absolute convergence tolerance. Default is 1e-9.}

\item{shock_vector}{a numeric vector indicating the stress shock (between 0 and 1) for each vertex.}
}
\value{
The \code{debt_rank} function returns an object of class \code{DebtRank}, which is a list containing:
\item{DebtRank}{a \code{data.frame} with the additional stress level and additional defaulted entities caused by each stressed vertex.}
\item{StressLevel}{a \code{data.frame} with the initial, final and additional stress level for each other vertex caused
by the stressed vertex.}

The \code{debt_rank_shock} function returns a object of class \code{DebtRankShock}, which is list containing:
\item{DebtRank}{a \code{data.frame} with the additional stress level and additional defaulted entities caused by the shock vector.}
\item{StressLevel}{a \code{data.frame} with the initial, final and additional stress level for each vertex caused
by the shock vector.}
}
\description{
The \code{debt_rank} function computes the DebtRank, additional defaults and the
differences in stress levels caused by each vertex of the network
when considering its default.

You can also use the auxiliary function \code{debt_rank_shock} to
simulate arbitrary shocks in the network.
}
\examples{
# Creating example data
## Assets Matrix (bilateral exposures)
assets_matrix <- matrix(c(0, 10, 3, 1, 0, 2, 0, 3, 0), ncol = 3)
rownames(assets_matrix) <- colnames(assets_matrix) <- letters[1:3]

## Capital Buffer
buffer <- c(a = 2, b = 5, c = 2)

## "Size" of the nodes
weights <-  c(a = 10, b = 100, c = 30)

# DebtRank - computes stress for each node considering its default
debt_rank(assets_matrix, buffer, weights)

# Arbitray shock -- 10\% stress shock in each node
shock <- c(a = 0.1, b = 0.1, c = 0.1)
debt_rank_shock(assets_matrix, buffer, shock, weights)

}
\references{
Battiston, S.; Puliga, M.; Kaushik, R.; Tasca, P.; Caldarelli, G. (2012).
DebtRank: Too central to fail? Financial Networks, the FED and systemic risk.
Scientific Reports, 2:541.
}

