% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DebtRank.R
\name{debt_rank_shock}
\alias{debt_rank}
\alias{debt_rank_shock}
\title{DebtRank of the vertices}
\usage{
debt_rank_shock(exposures, capital_buffer, weights, shock_vector,
  binary = FALSE, exposure_type = c("assets", "liabilities",
  "vulnerability"), max.it = 100, abs.tol = 1e-09)

debt_rank(exposures, capital_buffer, weights, binary = FALSE,
  exposure_type = c("assets", "liabilities", "vulnerability"), max.it = 100,
  abs.tol = 1e-09)
}
\arguments{
\item{exposures}{an adjacency \code{\link{matrix}}, (sparse) \code{\link{Matrix}} or an \code{\link[igraph]{igraph}}
object with the network of bilateral exposures between vertices. By default, the function
expects the exposures in the form of an assets matrix
in which A -> B means that A has an asset with B. However, you can
change that with the parameter \code{exposure_type}. When using a matrix, preferably it should have
rows and columns names.}

\item{capital_buffer}{a numeric vector with the capital buffer for each vertex.
Values should be in the same row/column order as the network of bilateral exposures. This
parameter is not needed if \code{exposure_type = "vulnerability"}.}

\item{weights}{This should be a vector
representing the relative importance of each vertex in the network.
If the vector does not sum to 1, it will be normalized. You should use
weights that have some practical significance like,
for instance, the total assets of the nodes in the case of a financial network.}

\item{shock_vector}{a numeric vector indicating the stress shock (between 0 and 1) for each vertex.}

\item{binary}{if \code{binary = TRUE} the function computes the binary vulnerability matrix,
truncating all values less than 1 to 0. Otherwise, 
it uses the partial stress version of the vulnerability matrix.}

\item{exposure_type}{character vector indicating the type of the bilateral exposures. It can be
an "assets" network (where A -> B means that A has an asset with B),
a "liabilities" network (where A -> B means that A has a debt with B) or
it can be a (binary) "vulnerability" matrix (where A -> B indicates the relative impact
 of A in B's capital buffer).For the last option see \code{\link{vulnerability_matrix}}.
 The default is "assets".}

\item{max.it}{the maximum number of iterations. Default is 100.}

\item{abs.tol}{the absolute convergence tolerance. Default is 1e-9.}
}
\value{
The \code{debt_rank} function returns an object of class \code{DebtRank}, 
which is a list containing:
\item{DebtRank}{a \code{data.frame} with the additional stress level and 
the number of additional defaulted entities caused 
by each stressed vertex (initial shock scenario).}
\item{StressLevel}{a \code{data.frame} with the initial, final and additional stress level for each other vertex caused
by the stressed vertex (initial shock scenario).}

The \code{debt_rank_shock} function returns an object of class \code{DebtRankShock}, 
which is list containing:
\item{DebtRank}{a \code{data.frame} with the additional stress level and 
the number of additional defaulted entities caused by the initial shock vector.}
\item{StressLevel}{a \code{data.frame} with the initial, final and additional stress level for each vertex caused
by the initial shock vector.}
}
\description{
The \code{debt_rank} function computes the DebtRank, 
the number of additional defaults and the
differences in stress levels caused by each vertex of the network
when we consider as the initial shock scenario its default.

You can also use the auxiliary function \code{debt_rank_shock} to
simulate arbitrary shocks in the network.
}
\examples{
# Creating example data
## Assets Matrix (bilateral exposures)
assets_matrix <- matrix(c(0, 10, 3, 1, 0, 2, 0, 3, 0), ncol = 3)
rownames(assets_matrix) <- colnames(assets_matrix) <- letters[1:3]

## Capital Buffer
buffer <- c(a = 2, b = 5, c = 2)

## "Size" of the nodes
weights <-  c(a = 10, b = 100, c = 30)

# DebtRank - computes stress for each node considering its default
debt_rank(exposures = assets_matrix, capital_buffer = buffer, weights = weights)

# Arbitrary shock -- 10\% stress shock for each node
shock <- c(a = 0.1, b = 0.1, c = 0.1)
debt_rank_shock(exposures = assets_matrix, 
                capital_buffer = buffer, 
                weights = weights, 
                shock_vector = shock)

}
\references{
Battiston, S.; Puliga, M.; Kaushik, R.; Tasca, P.; Caldarelli, G. (2012).
DebtRank: Too central to fail? Financial Networks, the FED and systemic risk.
Scientific Reports, 2:541.
}

