% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerability_communicability.R
\name{vulnerability_matrix}
\alias{vulnerability_matrix}
\title{Computes the (binary) vulnerability matrix}
\usage{
vulnerability_matrix(exposures, capital_buffer, binary = TRUE,
  exposure_type = c("assets", "liabilities", "vulnerability"))
}
\arguments{
\item{exposures}{an adjacency \code{\link{matrix}} or an \code{\link[igraph]{igraph}}
object with the network of bilateral exposures between vertices. By default, the function
expects the exposures in the form of an assets matrix
in which A -> B means that A has an asset with B. However, you can
change that with the parameter \code{exposure_type}. When using a matrix, preferably it should have
rows and columns names.}

\item{capital_buffer}{a numeric vector with the capital buffer for each vertex.
Values should be in the same row/column order as the network of bilateral exposures. This
parameter is not needed if \code{exposure_type = "vulnerability"}.}

\item{binary}{if \code{binary = TRUE} it computes the binary vulnerability matrix,
truncating all values less than 1 to 0.}

\item{exposure_type}{character vector indicating the type of the bilateral exposures. It can be
an "assets" network (where A -> B means that A has an asset with B),
a "liabilities" network (where A -> B means that A has a debt with B) or
it can be a (binary) "vulnerability" matrix (where A -> B indicates the relative impact
 of A in B's capital buffer).For the last option see \code{\link{vulnerability_matrix}}.
 The default is "assets".}
}
\value{
The function returns a (binary) vulnerability matrix.

The term V[i,j] of the vulnerability matrix represents
the impact of i's default in j's capital buffer. If \code{binary = TRUE}
the values less than 1 are truncated to zero.
}
\description{
The function computes a vulnerability matrix given
a network of bilateral exposures and a vector of capital buffers. The vulnerability
matrix represents how much a vertex impacts the capital buffer of another vertex
if it defaults.
}
\examples{
# Creating example data
## Assets Matrix (bilateral exposures)
assets_matrix <- matrix(c(0, 10, 3, 1, 0, 2, 0, 3, 0), ncol = 3)
rownames(assets_matrix) <- colnames(assets_matrix) <- letters[1:3]

## Capital Buffer
buffer <- c(a = 2, b = 5, c = 2)

# Vulnerability matrices
vulnerability_matrix(assets_matrix, buffer, binary = FALSE)
vulnerability_matrix(assets_matrix, buffer, binary = TRUE)
}

